refactoring coupons and orders tests
